Implemented pool allocator in STGen
Tracking unique address ranges with a <multiset> cost
too many mallocs/frees with the default allocator.
A custom, but simplistic block pool allocator was implemented
and improves performance by 1.25x specifically with blackscholes.
Performance is also improved with other benchmarks, but is not
explicitly measured.

Some minor cleanup with the dummy backend merged into this commit.