Added convenience features for users
Enabled sigil2 to be put in or run from other directories. Internal
paths for sigrind are now based on the location of the 'sigil2' binary
instead of the current working directory.

The wrapper library for sigrind is now required to be in the same
directory as the 'sigil2' binary. If it is in a different directory
and LD_PRELOAD is set, 'sigil2' will still detect sync events, but
will falsely say the wrapper is not found.

SynchroTraceGen still outputs only to the current working directory.