Force valgrind to run threads in round-robin
Valgrind is now forcibly forked with --fair-sched=yes option.
This option forces each thread, in a multithreaded application,
to be serialized in a round robin fashion. There is little reason
to leave it at the default, in which one thread may dominate and
starve other threads, resulting in less reproducible event generation.