Packed event structs for memory optimization
External testing has shown that, at least on x86 machines, packed structs
vs padded structs do not have a significant difference in performance.
Additionally, because the structs are streamed in a memory buffer,
any additional cache misses will need to be read in sequentially, anyway.

The reduction in memory usage should be more significant in this case,
than the reduction in performance. Small tests showed a reduction of
around 2GB, and a performance reduction of around 2%.