Add special capnproto uncompressed format
Uncompressed events only contain one memory read/write.
The compressed events (split-by-N) can have multiple memory accesses,
and require lists of addresses in the CapnProto schema.

The lists are very inefficient in the case of an uncompressed trace.
Uncompressed traces are desirable in situations where we want full
granularity so it's not an uncommon scenario.

This adds an additional capnproto format for uncompressed traces,
which shows nearly an order of magnitude increase in space efficiency
and trace generation.

This feature required additional logger logic that's specialized for
the case of an uncompressed trace, so a 'text' version of the
uncompressed logger was also implemented, however there should be no
change in the text format.