VG submodule now points to VG official git repo
Sigrind (Valgrind Sigil2 client) is now a git patch to the main
Valgrind repo, instead of a custom forked repo.

The client is self-contained enough to warrant removing the
dependency and requirement of the forked repo.

This will make maintenance easier, at the cost of depending on
an external project.