Update STGen text logger to shortened format
With "-ltextv2" option passed to stgen, a better format
for parsing will be used.
(no need to include thread/event ids on each line,
 and unique identifiers for each event type)

Update STGen to use fmtlib vs custom formatting.

spdlog: update to v1.3.1

The change to fmtlib combined with the spdlog async changes
causes a performance degradation (~10% in a toy benchmark).
This is acceptable, at the moment, for the improvements in
maintainability, as I am looking at this code less often and
do not plan to be heavily involved in its continuation.