Add support for variable-sized event ser/de
Adds events that vary in size depending on event-type and event
configuration, for added zoom-in/zoom-out support.
Some event metadata is bit-packed, but bit-packed data never crosses
byte boundaries, and is always at least byte-aligned.

Prior implementation used a large union for all types, which wasted
a lot of buffer space. This change increases the complexity of generating
and parsing events, with the goal of greater flexibility, less bandwidth,
and lower memory footprint.