Add Pytorch-Caffe2 frontend support
Existing support only expects limited net and operator events.

--executable=ARGS is expected to be a python script name, with args to
the script. The internal prism wrapper then invokes 'python' with ARGS.

The multiple layers involved in PyTorch/C2 requires some
configurability constraints. The user script is expected to manually
initialize Caffe2 with appropriate Prism args, attach the PrismObserver,
and run the net appropriately.

This is in contrast to the possibility of just providing an offline
model spec, and performing all the setup for the user.
It is a trade-off in complexity and configurability.